change type of field price

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interaction_api/model/dto/ProductDto.java
@@ -1,11 +1,12 @@
 package ru.yandex.practicum.interaction_api.model.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 import ru.yandex.practicum.interaction_api.enums.ProductCategory;
 import ru.yandex.practicum.interaction_api.enums.ProductState;
 import ru.yandex.practicum.interaction_api.enums.QuantityState;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -17,5 +18,5 @@ public class ProductDto {
     private QuantityState quantityState;
     private ProductState productState;
     private ProductCategory productCategory;
-    private Integer price;
+    private BigDecimal price;
 }
